Fix bugs with file and spreadsheet selection
The selection of Windows backup files did not account for the cases where the identifier is not at the beginning of the file path.

Also, when we do not find a table with headers that are sought for, slicing failed because we did not previously condition for this event.

Finally, the call for command line arguments will only occur in a non-interactive session.