Bug successfully fixed
The problem lied with the way purrr::map and its sisters were used. It was therefore decided that the apply() family of functions would be easier to use (and debug). The key change that was made was in the 'simplify = FALSE' argument passed to the inner sapply function.

Also, the code for creating the consolidated data frame was packaged into a function named 'fix_multip'.

So there.